pyproject.toml
[tool.poetry]
name = "Chuck"
version = "0.1.0"
description = ""
authors = ["ps4rtnite1234@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.0"
aiohttp = "^3.7.4"
discord = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
